refactor: extract row iteration utilities into row_iter.go
Move generic row/iterator utilities from execute_sql.go to row_iter.go:
- extractColumnNames, parseQueryStats, consumeRowIter*, spannerRowToRow

These are cross-cutting utilities used by multiple files, not specific
to SQL query execution. This makes execute_sql.go purely about the
query execution pipeline (368 lines).
